remove dtype from rms class initialization

+++ src/tiny_sota/models/qwen_arch.py
@@ -11,8 +11,8 @@ class DecoderBlock(nn.Module):
         super(DecoderBlock,self).__init__()
         emb_dim = config.emb_dim 
         dtype = config.dtype
-        self.rms1 = RMSNorm(emb_dim, is_qwen3=is_qwen3, dtype=dtype)
-        self.rms2 = RMSNorm(emb_dim, is_qwen3=is_qwen3, dtype=dtype)
+        self.rms1 = RMSNorm(emb_dim, is_qwen3=is_qwen3)
+        self.rms2 = RMSNorm(emb_dim, is_qwen3=is_qwen3)
         self.attn = GQAttention(config, is_qwen3=True)
         self.feed_forward = TriFeedForward(config)
     def forward(self, x, mask, cos, sin):
@@ -38,7 +38,7 @@ class Qwen3Model(nn.Module):
         self.embedding = nn.Embedding(vocab, emb_dim, dtype=dtype)
         self.decoders = nn.ModuleList([
             DecoderBlock(config) for _ in range(layers)])
-        self.rms_norm = RMSNorm(emb_dim, is_qwen3=is_qwen3, dtype=dtype)
+        self.rms_norm = RMSNorm(emb_dim, is_qwen3=is_qwen3)
         self.linear = nn.Linear(emb_dim, vocab, bias=bias, dtype=dtype)
         mask = torch.triu(torch.ones(context_len,context_len), diagonal=1)
         cos, sin = compute_rope_params(head_dim, context_len, config.rope_base)
